add dom and set state
