+++ renderProductTesting.js
@@ -1,27 +1,29 @@
 //note: switch up/chnage render function to hard code radio buttons and then use query selector to loop through radio buttons
 
 
-export default function renderProductTesting(productData) {
+//This function will go away :( Commenting out for now in case I can reuse any of the pieces. 
 
-    //get product id from productData.js
-    const getProductId = document.getElementById(productData.id);
+// export default function renderProductTesting(productData) {
+
+//     //get product id from productData.js
+//     const getProductId = document.getElementById(productData.id);
 
 
-    //create a radio input button for user selection
-    const radioButton = document.createElement ('input');
-    radioButton.setAttribute('type', 'radio');
+//     //create a radio input button for user selection
+//     const radioButton = document.createElement ('input');
+//     radioButton.setAttribute('type', 'radio');
 
-    //create an image for the radio button
-    const img = document.createElement('img'); 
-    img.src = '../assets/' + productData.image;
-    radioButton.appendChild(img);
+//     //create an image for the radio button
+//     const img = document.createElement('img'); 
+//     img.src = '../assets/' + productData.image;
+//     radioButton.appendChild(img);
 
-    //create a p element to display the name of the product with the image
-    const productNameP = document.createElement('p');
-    productNameP.textContent = productData.name;
-    radioButton.appendChild(productNameP);
+//     //create a p element to display the name of the product with the image
+//     const productNameP = document.createElement('p');
+//     productNameP.textContent = productData.name;
+//     radioButton.appendChild(productNameP);
 
-    return radioButton;
-}
+//     return radioButton;
+// }
